# CSV/TSV use a pool of buffered readers to avoid allocs